requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
cfgv==3.4.0
distlib==0.3.8
fastapi==0.114.1
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.0
idna==3.8
iniconfig==2.0.0
mypy==1.11.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.6.4
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.22
starlette==0.38.5
typing_extensions==4.12.2
virtualenv==20.26.3
